Support regular expression matching
If the value of the target domain field ends with a $ it is treated as
a regular expression. The $ is part of the expression, which makes
sure using the match length to find the most precise matching rule
still works. If there are multiple matches with the same length, the
first rule from top is used.

Note that the regexp is used as-is, you need to consider the start of
the URL and subdomain boundaries yourself. For example, the regexp
"example\.com$" would also match "my-example.com". To match only
example.com and its subdomains with a regular expression, use
"(^|\.)example\.com$".

Closes #6.